\name{calculate_prior_errors}
\alias{calculate_prior_errors}
\title{calculate_prior_errors}
\usage{
calculate_prior_errors(inFile_prefix, save=F)
}
\description{

}
\arguments{
\item{inFile_prefix}{Character. Prefix to file to analyse, named inFile_prefix+"_countsPNQ.txt"."}
\item{save}{Logical. Should results be saved into file named: inFile_prefix+"_prior_errors.txt" .}
\item{intercep}{Numeric. Only one value.}
}
\details{
Compute prior probabilities or errors analysing the number of errors made during sequencing a reference strand of nucleotides.
}
\examples{

}

