\name{SplitSequencesInFiles}
\alias{SplitSequencesInFiles}
\title{SplitSequencesInFiles}
\usage{
SplitSequencesInFiles(inFile_prefix, reference_sequence, columns=NULL)
}
\description{
Creates one file per sequence analyzed.
}
\arguments{
  \item{inFile_prefix}{Character. Name of the file to analyse, without the ".maf".}
  \item{reference_sequence}{Character vector. Reference sequence in a vector (one nucleotide one element of the vector, in order.)}
\item{columns}{Numeric. Positions to be analyzed. If NULL, all of them.}
}
\details{

}
\examples{
}

